Struct stm32f103xx_hal::dma::CircBuffer

pub struct CircBuffer<BUFFER, CHANNEL, RESOURCE> where
    BUFFER: 'static,  { /* fields omitted */ }

Methods

impl<B: AsRef<[u8]>, RE> CircBuffer<B, C1, RE>

pub fn peek_whole_buffer<R, F>(&mut self, f: F) -> Result<R, Error> where
F: FnOnce(&B, &B, usize) -> R,

peers into the buffer without touching any flag

pub fn retrieve_filled_half<R, F>(&mut self, f: F) -> Result<R, Error> where
F: FnOnce(&B, Half, usize) -> R,

Give access to the available half buffer this also flags it as clear to re-fill.

pub fn get_filled_half(&mut self) -> Result<Half, Error>

Returns the Half of the buffer that can be read this also flags it as clear to re-fill.

pub fn release(self) -> (&'static mut [B; 2], C1, RE)

impl<B: AsRef<[u8]>, RE> CircBuffer<B, C2, RE>

pub fn peek_whole_buffer<R, F>(&mut self, f: F) -> Result<R, Error> where
F: FnOnce(&B, &B, usize) -> R,

peers into the buffer without touching any flag

pub fn retrieve_filled_half<R, F>(&mut self, f: F) -> Result<R, Error> where
F: FnOnce(&B, Half, usize) -> R,

Give access to the available half buffer this also flags it as clear to re-fill.

pub fn get_filled_half(&mut self) -> Result<Half, Error>

Returns the Half of the buffer that can be read this also flags it as clear to re-fill.

pub fn release(self) -> (&'static mut [B; 2], C2, RE)

impl<B: AsRef<[u8]>, RE> CircBuffer<B, C3, RE>

pub fn peek_whole_buffer<R, F>(&mut self, f: F) -> Result<R, Error> where
F: FnOnce(&B, &B, usize) -> R,

peers into the buffer without touching any flag

pub fn retrieve_filled_half<R, F>(&mut self, f: F) -> Result<R, Error> where
F: FnOnce(&B, Half, usize) -> R,

Give access to the available half buffer this also flags it as clear to re-fill.

pub fn get_filled_half(&mut self) -> Result<Half, Error>

Returns the Half of the buffer that can be read this also flags it as clear to re-fill.

pub fn release(self) -> (&'static mut [B; 2], C3, RE)

impl<B: AsRef<[u8]>, RE> CircBuffer<B, C4, RE>

pub fn peek_whole_buffer<R, F>(&mut self, f: F) -> Result<R, Error> where
F: FnOnce(&B, &B, usize) -> R,

peers into the buffer without touching any flag

pub fn retrieve_filled_half<R, F>(&mut self, f: F) -> Result<R, Error> where
F: FnOnce(&B, Half, usize) -> R,

Give access to the available half buffer this also flags it as clear to re-fill.

pub fn get_filled_half(&mut self) -> Result<Half, Error>

Returns the Half of the buffer that can be read this also flags it as clear to re-fill.

pub fn release(self) -> (&'static mut [B; 2], C4, RE)

impl<B: AsRef<[u8]>, RE> CircBuffer<B, C5, RE>

pub fn peek_whole_buffer<R, F>(&mut self, f: F) -> Result<R, Error> where
F: FnOnce(&B, &B, usize) -> R,

peers into the buffer without touching any flag

pub fn retrieve_filled_half<R, F>(&mut self, f: F) -> Result<R, Error> where
F: FnOnce(&B, Half, usize) -> R,

Give access to the available half buffer this also flags it as clear to re-fill.

pub fn get_filled_half(&mut self) -> Result<Half, Error>

Returns the Half of the buffer that can be read this also flags it as clear to re-fill.

pub fn release(self) -> (&'static mut [B; 2], C5, RE)

impl<B: AsRef<[u8]>, RE> CircBuffer<B, C6, RE>

pub fn peek_whole_buffer<R, F>(&mut self, f: F) -> Result<R, Error> where
F: FnOnce(&B, &B, usize) -> R,

peers into the buffer without touching any flag

pub fn retrieve_filled_half<R, F>(&mut self, f: F) -> Result<R, Error> where
F: FnOnce(&B, Half, usize) -> R,

Give access to the available half buffer this also flags it as clear to re-fill.

pub fn get_filled_half(&mut self) -> Result<Half, Error>

Returns the Half of the buffer that can be read this also flags it as clear to re-fill.

pub fn release(self) -> (&'static mut [B; 2], C6, RE)

impl<B: AsRef<[u8]>, RE> CircBuffer<B, C7, RE>

pub fn peek_whole_buffer<R, F>(&mut self, f: F) -> Result<R, Error> where
F: FnOnce(&B, &B, usize) -> R,

peers into the buffer without touching any flag

pub fn retrieve_filled_half<R, F>(&mut self, f: F) -> Result<R, Error> where
F: FnOnce(&B, Half, usize) -> R,

Give access to the available half buffer this also flags it as clear to re-fill.

pub fn get_filled_half(&mut self) -> Result<Half, Error>

Returns the Half of the buffer that can be read this also flags it as clear to re-fill.

pub fn release(self) -> (&'static mut [B; 2], C7, RE)

impl<B: AsRef<[u8]>, RE> CircBuffer<B, C1, RE>

pub fn peek_whole_buffer<R, F>(&mut self, f: F) -> Result<R, Error> where
F: FnOnce(&B, &B, usize) -> R,

peers into the buffer without touching any flag

pub fn retrieve_filled_half<R, F>(&mut self, f: F) -> Result<R, Error> where
F: FnOnce(&B, Half, usize) -> R,

Give access to the available half buffer this also flags it as clear to re-fill.

pub fn get_filled_half(&mut self) -> Result<Half, Error>

Returns the Half of the buffer that can be read this also flags it as clear to re-fill.

pub fn release(self) -> (&'static mut [B; 2], C1, RE)

impl<B: AsRef<[u8]>, RE> CircBuffer<B, C2, RE>

pub fn peek_whole_buffer<R, F>(&mut self, f: F) -> Result<R, Error> where
F: FnOnce(&B, &B, usize) -> R,

peers into the buffer without touching any flag

pub fn retrieve_filled_half<R, F>(&mut self, f: F) -> Result<R, Error> where
F: FnOnce(&B, Half, usize) -> R,

Give access to the available half buffer this also flags it as clear to re-fill.

pub fn get_filled_half(&mut self) -> Result<Half, Error>

Returns the Half of the buffer that can be read this also flags it as clear to re-fill.

pub fn release(self) -> (&'static mut [B; 2], C2, RE)

impl<B: AsRef<[u8]>, RE> CircBuffer<B, C3, RE>

pub fn peek_whole_buffer<R, F>(&mut self, f: F) -> Result<R, Error> where
F: FnOnce(&B, &B, usize) -> R,

peers into the buffer without touching any flag

pub fn retrieve_filled_half<R, F>(&mut self, f: F) -> Result<R, Error> where
F: FnOnce(&B, Half, usize) -> R,

Give access to the available half buffer this also flags it as clear to re-fill.

pub fn get_filled_half(&mut self) -> Result<Half, Error>

Returns the Half of the buffer that can be read this also flags it as clear to re-fill.

pub fn release(self) -> (&'static mut [B; 2], C3, RE)

impl<B: AsRef<[u8]>, RE> CircBuffer<B, C4, RE>

pub fn peek_whole_buffer<R, F>(&mut self, f: F) -> Result<R, Error> where
F: FnOnce(&B, &B, usize) -> R,

peers into the buffer without touching any flag

pub fn retrieve_filled_half<R, F>(&mut self, f: F) -> Result<R, Error> where
F: FnOnce(&B, Half, usize) -> R,

Give access to the available half buffer this also flags it as clear to re-fill.

pub fn get_filled_half(&mut self) -> Result<Half, Error>

Returns the Half of the buffer that can be read this also flags it as clear to re-fill.

pub fn release(self) -> (&'static mut [B; 2], C4, RE)

impl<B: AsRef<[u8]>, RE> CircBuffer<B, C5, RE>

pub fn peek_whole_buffer<R, F>(&mut self, f: F) -> Result<R, Error> where
F: FnOnce(&B, &B, usize) -> R,

peers into the buffer without touching any flag

pub fn retrieve_filled_half<R, F>(&mut self, f: F) -> Result<R, Error> where
F: FnOnce(&B, Half, usize) -> R,

Give access to the available half buffer this also flags it as clear to re-fill.

pub fn get_filled_half(&mut self) -> Result<Half, Error>

Returns the Half of the buffer that can be read this also flags it as clear to re-fill.

pub fn release(self) -> (&'static mut [B; 2], C5, RE)