Struct stm32f103xx_hal::gpio::gpiob::PB13

pub struct PB13<MODE> { /* fields omitted */ }

Pin

Methods

impl<MODE> PB13<MODE>

pub fn into_alternate_push_pull(self, cr: &mut CRH) -> PB13<Alternate<PushPull>>

Configures the pin to operate as an alternate function push pull output pin

pub fn into_floating_input(self, cr: &mut CRH) -> PB13<Input<Floating>>

Configures the pin to operate as a floating input pin

pub fn into_push_pull_output(self, cr: &mut CRH) -> PB13<Output<PushPull>>

Configures the pin to operate as an push pull output pin

impl<MODE> PB13<Output<MODE>>

pub fn downgrade(self) -> PBx<Output<MODE>>

Erases the pin number from the type

This is useful when you want to collect the pins into an array where you need all the elements to have the same type