Struct stm32f103xx_hal::dma::CircBuffer

pub struct CircBuffer<BUFFER, CHANNEL> where
    BUFFER: 'static,  { /* fields omitted */ }

Methods

impl<B> CircBuffer<B, C1>

pub fn peek<R, F>(&mut self, f: F) -> Result<R, Error> where
F: FnOnce(&B, Half) -> R,

Peeks into the readable half of the buffer

pub fn readable_half(&mut self) -> Result<Half, Error>

Returns the Half of the buffer that can be read

impl<B> CircBuffer<B, C2>

pub fn peek<R, F>(&mut self, f: F) -> Result<R, Error> where
F: FnOnce(&B, Half) -> R,

Peeks into the readable half of the buffer

pub fn readable_half(&mut self) -> Result<Half, Error>

Returns the Half of the buffer that can be read

impl<B> CircBuffer<B, C3>

pub fn peek<R, F>(&mut self, f: F) -> Result<R, Error> where
F: FnOnce(&B, Half) -> R,

Peeks into the readable half of the buffer

pub fn readable_half(&mut self) -> Result<Half, Error>

Returns the Half of the buffer that can be read

impl<B> CircBuffer<B, C4>

pub fn peek<R, F>(&mut self, f: F) -> Result<R, Error> where
F: FnOnce(&B, Half) -> R,

Peeks into the readable half of the buffer

pub fn readable_half(&mut self) -> Result<Half, Error>

Returns the Half of the buffer that can be read

impl<B> CircBuffer<B, C5>

pub fn peek<R, F>(&mut self, f: F) -> Result<R, Error> where
F: FnOnce(&B, Half) -> R,

Peeks into the readable half of the buffer

pub fn readable_half(&mut self) -> Result<Half, Error>

Returns the Half of the buffer that can be read

impl<B> CircBuffer<B, C6>

pub fn peek<R, F>(&mut self, f: F) -> Result<R, Error> where
F: FnOnce(&B, Half) -> R,

Peeks into the readable half of the buffer

pub fn readable_half(&mut self) -> Result<Half, Error>

Returns the Half of the buffer that can be read

impl<B> CircBuffer<B, C7>

pub fn peek<R, F>(&mut self, f: F) -> Result<R, Error> where
F: FnOnce(&B, Half) -> R,

Peeks into the readable half of the buffer

pub fn readable_half(&mut self) -> Result<Half, Error>

Returns the Half of the buffer that can be read

impl<B> CircBuffer<B, C1>

pub fn peek<R, F>(&mut self, f: F) -> Result<R, Error> where
F: FnOnce(&B, Half) -> R,

Peeks into the readable half of the buffer

pub fn readable_half(&mut self) -> Result<Half, Error>

Returns the Half of the buffer that can be read

impl<B> CircBuffer<B, C2>

pub fn peek<R, F>(&mut self, f: F) -> Result<R, Error> where
F: FnOnce(&B, Half) -> R,

Peeks into the readable half of the buffer

pub fn readable_half(&mut self) -> Result<Half, Error>

Returns the Half of the buffer that can be read

impl<B> CircBuffer<B, C3>

pub fn peek<R, F>(&mut self, f: F) -> Result<R, Error> where
F: FnOnce(&B, Half) -> R,

Peeks into the readable half of the buffer

pub fn readable_half(&mut self) -> Result<Half, Error>

Returns the Half of the buffer that can be read

impl<B> CircBuffer<B, C4>

pub fn peek<R, F>(&mut self, f: F) -> Result<R, Error> where
F: FnOnce(&B, Half) -> R,

Peeks into the readable half of the buffer

pub fn readable_half(&mut self) -> Result<Half, Error>

Returns the Half of the buffer that can be read

impl<B> CircBuffer<B, C5>

pub fn peek<R, F>(&mut self, f: F) -> Result<R, Error> where
F: FnOnce(&B, Half) -> R,

Peeks into the readable half of the buffer

pub fn readable_half(&mut self) -> Result<Half, Error>

Returns the Half of the buffer that can be read