Trait stm32f103xx_hal::prelude::_embedded_hal_digital_StatefulOutputPin

pub trait _embedded_hal_digital_StatefulOutputPin {
    fn is_set_high(&self) -> bool;

    fn is_set_low(&self) -> bool;
}

Push-pull output pin that can read its output state

This trait is available if embedded-hal is built with the "unproven" feature.

Required Methods

fn is_set_high(&self) -> bool

Is the pin in drive high mode?

NOTE this does not read the electrical state of the pin

fn is_set_low(&self) -> bool

Is the pin in drive low mode?

NOTE this does not read the electrical state of the pin